+++ src/dataprotection/azext_dataprotection/tests/utils.py
@@ -8,18 +8,28 @@
 import time
 
 
-def wait_for_job_exclusivity_on_datasource(test, data_source_id=None):
+def wait_for_job_exclusivity_on_datasource(test, data_source_id=None, source_data_store=None):
     """ Checks if any job is in progress for pre-set datasource type and
         waits for its completion. Requires dataSourceType and dataSourceId kwargs.
     """
     if data_source_id:
         test.kwargs.update({'dataSourceId': data_source_id})
-    if not test.kwargs.get('dataSourceId'):
+    if source_data_store:
+        test.kwargs.update({'sourceDataStore': source_data_store})
+    if not test.kwargs.get('dataSourceId') or not test.kwargs.get('sourceDataStore'):
         return
-    jobs_in_progress = test.cmd('az dataprotection job list-from-resourcegraph --datasource-type "{dataSourceType}" --datasource-id "{dataSourceId}" --status "InProgress"').get_output_in_json()
+    jobs_in_progress = test.cmd('az dataprotection job list-from-resourcegraph '
+                                '--datasource-type "{dataSourceType}" '
+                                '--datasource-id "{dataSourceId}" '
+                                '--status "InProgress" '
+                                '--query "[?properties.sourceDataStoreType == \'{sourceDataStore}\']"').get_output_in_json()
     while jobs_in_progress:
         time.sleep(10)
-        jobs_in_progress = test.cmd('az dataprotection job list-from-resourcegraph --datasource-type "{dataSourceType}" --datasource-id "{dataSourceId}" --status "InProgress"').get_output_in_json()
+        jobs_in_progress = test.cmd('az dataprotection job list-from-resourcegraph '
+                                    '--datasource-type "{dataSourceType}" '
+                                    '--datasource-id "{dataSourceId}" '
+                                    '--status "InProgress" '
+                                    '--query "[?properties.sourceDataStoreType == \'{sourceDataStore}\']"').get_output_in_json()
 
 
 def track_job_to_completion(test, job_id=None):